requirements.txt
openai==1.83.0
langgraph
langchain-core
langchain-openai
python-dotenv